website and derived metric computations and searching are fully functional
