+++ tacc_stats/analysis/gen/jid_table.py
@@ -1,15 +1,24 @@
 from pandas import read_sql
+import time
+import psycopg2
+import tacc_stats.cfg as cfg
 
 class jid_table:
 
     def __init__(self, jid):
+        CONNECTION = "dbname=ls6_db1 host=localhost user=postgres port=5432"
+
         print("Initializing table for job {0}".format(jid))
 
         self.jid = jid
+
+        # Open temporary connection
+        self.conj = psycopg2.connect(CONNECTION)
+
         # Get job accounting data
-        acct_data = read_sql("""select * from job_data where jid = '{0}'""".format(jid), conn)
+        acct_data = read_sql("""select * from job_data where jid = '{0}'""".format(jid), self.conj)
         # job_data accounting host names must be converted to fqdn
-        acct_host_list = [h + '.' + cfg.host_name_ext for h in acct_data["host_list"].values[0]]
+        self.acct_host_list = [h + '.' + cfg.host_name_ext for h in acct_data["host_list"].values[0]]
     
         self.start_time = acct_data["start_time"].dt.tz_convert('US/Central').dt.tz_localize(None).values[0]
         self.end_time = acct_data["end_time"].dt.tz_convert('US/Central').dt.tz_localize(None).values[0]
@@ -18,8 +27,6 @@ class jid_table:
         qtime = time.time()
         sql = """drop table if exists job_{0}; select * into temp job_{0} from host_data where time between '{1}' and '{2}' and jid = '{0}'""".format(jid, self.start_time, self.end_time)
 
-        # Open temporary connection
-        self.conj = psycopg2.connect(CONNECTION)
         with self.conj.cursor() as cur:
             cur.execute(sql)
         print("query time: {0:.1f}".format(time.time()-qtime))
