+++ tacc_stats/site/machine/models.py
@@ -4,27 +4,28 @@ from django.db import models
 from django.forms import ModelForm
 from django.contrib.postgres.fields import ArrayField
 
-class job_data(models.Model):
-    class Meta:
-        db_table = 'job_data'
-        indexes = [models.Index(fields=['jid'])]
-        ordering = ['jid']
+# manage.py inspectdb
 
-    jid         = models.CharField(primary_key = True, max_length=32)
-    account     = models.CharField(max_length=64, null = True)
+class job_data(models.Model):
+    jid = models.CharField(primary_key = True, max_length=32)
     submit_time = models.DateTimeField()
-    start_time  = models.DateTimeField()
-    end_time    = models.DateTimeField(null = True)
-    runtime     = models.FloatField()
-    timelimit   = models.FloatField()
-    node_hrs    = models.FloatField()    
-    nhosts      = models.PositiveIntegerField(null = True)
-    ncores      = models.PositiveIntegerField(null = True)
-    username    = models.CharField(max_length = 64)
-    state       = models.CharField(max_length = 64)
-    queue       = models.CharField(max_length = 64)
-    jobname     = models.TextField()
+    start_time = models.DateTimeField()
+    end_time = models.DateTimeField()
+    runtime = models.FloatField(blank=True, null=True)
+    timelimit = models.FloatField(blank=True, null=True)
+    node_hrs = models.FloatField(blank=True, null=True)
+    nhosts = models.IntegerField(blank=True, null=True)
+    ncores = models.IntegerField(blank=True, null=True)
+    username = models.CharField(max_length=64)
+    account = models.CharField(max_length=64, blank=True, null=True)
+    queue = models.CharField(max_length=64, blank=True, null=True)
+    state = models.CharField(max_length=64, blank=True, null=True)
+    jobname = models.TextField(blank=True, null=True)
     host_list   = ArrayField(models.TextField())
+
+    class Meta:
+        managed = False
+        db_table = 'job_data'
     
     def __unicode__(self):
         return str(self.id)
@@ -37,44 +38,18 @@ class job_data(models.Model):
         else:
             ret_val = "silver"
         return ret_val
-"""
-class Proc(models.Model):
-    job  = models.ForeignKey(Job, on_delete = models.CASCADE)
-    name = models.CharField(max_length=128)
-    host = models.CharField(max_length=128)
-    uid  = models.IntegerField(null=True)
-    pid  = models.IntegerField(null=True)
-    VmHWM = models.FloatField(null=True)
-    Threads = models.IntegerField(null=True)
-    Cpus_allowed_list = models.CharField(max_length=128)
-    Mems_allowed_list = models.CharField(max_length=128)
 
-    def __unicode__(self):
-        return str(self.name)
+class metrics_data(models.Model):
+    jid = models.ForeignKey(job_data, on_delete = models.CASCADE, db_column='jid', blank=True, null=True)
+    type = models.CharField(max_length=32, blank=True, null=True)
+    metric = models.CharField(max_length=32, blank=True, null=True)
+    units = models.CharField(max_length=16, blank=True, null=True)
+    value = models.FloatField(blank=True, null=True)
 
-
-class Host(models.Model):
-    name = models.CharField(max_length=128)
-    jobs = models.ManyToManyField(Job)
-    class Meta: ordering = ('name',)
-
-    def __unicode__(self):
-        return str(self.name)
-
-class Libraries(models.Model):
-    object_path = models.CharField(max_length=1024)
-    module_name = models.CharField(max_length=64)
-    jobs = models.ManyToManyField(Job)
-
-    class Meta: ordering = ('object_path',)
-
-    def __unicode__(self):
-        return str(self.object_path)
-
-
-class JobForm(ModelForm):
     class Meta:
-        model = Job
-        fields = ['id']
+        managed = False
+        db_table = 'metrics_data'
+        unique_together = (('jid', 'type', 'metric'),)
 
-"""
+    def __unicode__(self):
+        return str(self.jid + '_' + type  + '_' + metric)
