+++ tacc_stats/site/machine/templates/machine/search.html
@@ -34,84 +34,108 @@ en: {
 {% endif %}
 <hr/>
 
-<form action="{% url 'machine:search' %}" method="get">
+<h4> Search fields are combined </h4>
+<form id="search" action="{% url 'machine:search' %}" method="get">
+<!-----{% csrf_token %}----->
+</br>
+  <div class="row">
+    <div class="col-md-1"><label>Start Date</label></div>
+    <div class="col-md-1"></div>
+    <div class="col-md-2"><input type="date" class="form-control input-sm" name="end_time__gte"></div>
+    <div class="col-md-1"></div>
+    <div class="col-md-1"><label>End Date</label></div>
+    <div class="col-md-2"><input type="date" class="form-control input-sm" name="end_time__lte"></div>
+  </div>
 
-<h4>Search for jobs on multiple fields </h4>
 </br>
-<div class="row">
-  <div class="col-md-1"><label>Start Date</label></div>
-  <div class="col-md-1"></div>
-  <div class="col-md-2"><input type="date" class="form-control input-sm" name="end_time__gte"></div>
-  <div class="col-md-1"></div>
-  <div class="col-md-1"><label>End Date</label></div>
-  <div class="col-md-2"><input type="date" class="form-control input-sm" name="end_time__lte"></div>
-</div>
-</br>
-<div class="row">
+
+  <div class="row">
     <div class="col-md-2"><label for="host">Host</label></div>
     <div class="col-md-2"><input type="text" class="form-control input-sm" name="host"></div>
-</div>
-<div class="row">
+  </div>
+
+</br>
+
+  <div class="row">
     <div class="col-md-2"><label for="username">Username</label></div>
     <div class="col-md-2"><input type="text" class="form-control input-sm" name="username"></div>
-</div>
-<div class="row">
+  </div>
+  <div class="row">
     <div class="col-md-2"><label for="account">Account</label></div>
     <div class="col-md-2"><input type="text" class="form-control input-sm" name="account__icontains"></div>
+  </div>
+
+
+  <div class="row">    
+    <div class="col-md-2"><label for="State">State</label></div>
+    <div class="col-md-2 btn-group" data-toggle="buttons" aria-label="State">
+      <select type="text" id="state" name="state">
+	{% for state in choice.state.field.choices %}
+	<option>{{ state.0 }}</option>
+	{% endfor %}
+      </select>
+    </div>
+  </div>
+
+  <div class="row">    
+    <div class="col-md-2"><label for="Queue">Queue</label></div>
+    <div class="col-md-2 btn-group" data-toggle="buttons" aria-label="Queue">
+      <select type="text" id="queue" name="queue">
+	{% for queue in choice.queue.field.choices %}
+	<option>{{ queue.0 }}</option>
+	{% endfor %}
+      </select>
+    </div>
+  </div>
+
+</br>
+
+<h4>Search on Resources</h4>
+
+<div class="row">
+    <div class="col-md-2"><label for="runtime">Runtime</label></div>
+    <div class="col-md-1"><input id="runtime" type="text" class="form-control input-sm" name="runtime__gte" placeholder="min seconds"></div>
+    <div class="col-md-1"><input id="runtime" type="text" class="form-control input-sm" name="runtime__lte" placeholder="max seconds"></div>
 </div>
 <div class="row">
-    <div class="col-md-2"><label for="uid">Executable</label></div>
-    <div class="col-md-2"><input type="text" class="form-control input-sm" name="exe__icontains"></div>
+    <div class="col-md-2"><label for="nodes">Nodes</label></div>
+    <div class="col-md-1"><input id="nodes" type="text" class="form-control input-sm" name="nhosts__gte" placeholder="min nodes"></div>
+    <div class="col-md-1"><input id="nodes" type="text" class="form-control input-sm" name="nhosts__lte" placeholder="max nodes"></div>
 </div>
 <div class="row">
-    <div class="col-md-2"><label for="uid">Process</label></div>
-    <div class="col-md-2"><input type="text" class="form-control input-sm" name="proc__name__icontains"></div>
+    <div class="col-md-2"><label for="runtime">Node-hrs</label></div>
+    <div class="col-md-1"><input type="text" class="form-control input-sm" name="node_hrs__gte" placeholder="min nodes-hrs"></div>
+    <div class="col-md-1"><input type="text" class="form-control input-sm" name="node_hrs__lte" placeholder="max node-hrs"></div>
 </div>
-<div class="row">
-    <div class="col-md-2"><label for="uid">Queue</label></div>
-    <div class="col-md-2"><input type="text" class="form-control input-sm" name="queue"></div>
-</div>
+
+
 <br/>
+
+
+ <h4>Search on Derived Metrics</h4>
+
 <div class="row">
-    <div class="col-md-2"><label for="uid">Search field</label></div>
-    <div class="col-md-2">
-        <input type="text" class="form-control input-sm" name="opt_field1">
-        <input type="text" class="form-control input-sm" name="opt_field2">
-        <input type="text" class="form-control input-sm" name="opt_field3">
+    {% for metric in metrics %}
+    <div class="col-md-2"><label for="{{ metric.metric }}">{{ metric.metric }}</label></div>
+    <div class="col-md-1">
+        <input id="{{ metric.metric }}" type="text" class="form-control input-sm" name="metrics_{{ metric.metric }}__gte" placeholder="Min {{ metric.units }}">
     </div>
-    <div class="col-md-1"></div>
-    <div class="col-md-1"><label for="uid">Value</label></div>
-    <div class="col-md-2">
-        <input type="text" class="form-control input-sm" name="value1">
-        <input type="text" class="form-control input-sm" name="value2">
-        <input type="text" class="form-control input-sm" name="value3">
+    <div class="col-md-1">
+        <input id="{{ metric.metric }}" type="text" class="form-control input-sm" name="metrics_{{ metric.metric }}__lte" placeholder="Max {{ metric.units }}">
     </div>
+    {% endfor %}
 </div>
 <br/>
 
-<button type="submit" class="btn btn-default">Search</button>
+<button id="submit" type="submit" class="btn btn-default">Search</button>
 </form>
-
 
 </div>
 <hr>
 
-<div class="table-responsive">
-  <table class="table table-condensed table-bordered" style="max-width=75%">
-    <thead><font size="5" color="#002868">Top 10 Lustre jobs in past 5 days: JobID(Open/Close IOPS)</font></thead>
-    {% for key,values in md_job_list.items %}
-    <tr><td><a href="{% url 'machine:user_view' key %}"> {{ key }}</a> </td>
-      {% for v,d in values %} 
-      <td><a href="{% url 'machine:job' v %}">{{ v }}</a>({{ d|stringformat:"0.1f" }})</td>
-      {% endfor %}
-    </tr>
-    {% endfor %}
-  </table>
-</div>
-
 <div class="container-fluid">
 
-<h4>View all jobs for a given date </h4>
+<h4>List all jobs for a given date </h4>
 {% if date_list %}
 
 <nav class="navbar navbar-default" role="navigation">
@@ -131,5 +155,20 @@ en: {
     <p> No job data available </p>
 {% endif %}   
 </div>
+
+
+<script>
+jQuery(document).ready(function($){
+  
+$("search").submit(function() {
+$(this).find(":input").filter(function(){ return !this.value; }).attr("disabled", "disabled");
+return true; // ensure form still submits
+});
+
+$( "search" ).find( ":input" ).prop( "disabled", false );
+
+}
+</script>
+
 
 {% endblock %}
